Subhadeep

See merge request cnl-projects/fy24-java-batch-16/team-7/pg-spring-boot!9

+++ src/main/java/com/virtusa/project/PgRental/model/UserFavorites.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class UserFavorites {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int favoriteId;
+    private Long favoriteId;
 
     @ManyToOne
     @JoinColumn(name = "userId")
